+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Split deterministic timer into monotonic/realtime (#474)
 * Users can change realtime (#474)
 * Users can set the initial monotonic time (#480)
+* New commandline option to run a lua script
 
 ### Changed
 ### Fixed